Fix memory leaks, threading UB, logic errors, and deprecated API usage
Memory leaks: SessionController dtor (mAnalyser, mIncomingPackets),
UserBlockManager dtor + removeBlockEntity, QMouseEvent stack-allocated,
throw-by-pointer in FileTransferManager.

Threading/lifecycle: async connect via error signal (no waitForConnected),
mTimer NULL-check, deleteLater in form_Main event handlers.

Logic/UB: mSilence const-assignment removed, comma operator in Core ctor,
saveUserList() const removed, mNextOnlineStatus initialized, default: cases
in switches, quint64/qint64 types for >2GB transfer ETA, redundant nested if.

Cleanup: unused ElideLabel, Qt::SkipEmptyParts, QFlags(0)->WindowFlags(),
QByteArray deprecated insert/operator+= fixes.

+++ src/backend/FileTransferReceive.cpp
@@ -379,10 +379,10 @@ void CFileTransferReceive::start(QString FilePath, bool Accepted) {
 }
 
 void CFileTransferReceive::slotCalcAverageTransferSpeed() {
-  int departedtime = (mTimer.elapsed() / 1000);
+  qint64 departedtime = (mTimer.elapsed() / 1000);
   if (departedtime <= 0)
     departedtime = 1;
-  int speed = mAlreadyReceivedSize / departedtime;
+  quint64 speed = mAlreadyReceivedSize / departedtime;
 
   QString speedSize;
   QString speedType;
@@ -400,9 +400,9 @@ void CFileTransferReceive::doConvertNumberToTransferSize(quint64 inNumber,
   return mCore.doConvertNumberToTransferSize(inNumber, outNumber, outType, addStoOutType);
 }
 
-void CFileTransferReceive::CalcETA(int speed) {
+void CFileTransferReceive::CalcETA(quint64 speed) {
   QString EmitString;
-  int secLeft;
+  quint64 secLeft;
 
   if (speed > 0) {
     secLeft = (mFileSize - mAlreadyReceivedSize) / speed;
